Create Array Concep-2
Zero Removal and Array Rearrangement

This simple JavaScript project takes a user-specified number of values as input, removes zeroes from the array, shifts the non-zero elements to the front, and fills the remaining positions with zeros. The program maintains the order of non-zero elements and logs the final array to the console.